chore: Tiny cleanup in mining

+++ src/stats/fishing.rs
@@ -242,7 +242,7 @@ impl IntoString for FishingDetails {
                 "{} {}",
                 c1(format!("{}:", d.name.as_str()).as_str()),
                 c2(common::commas_from_string(
-                    format!("{}", (xp_difference / (self.xp as f64 * d.value)).ceil()).as_str(),
+                    format!("{}", (xp_difference / (self.xp * d.value)).ceil()).as_str(),
                     "d"
                 )
                 .as_str())
